requirements.txt
fastapi
uvicorn
celery
redis
kafka-python
